Refactor code from init.py
